Add PinchTracker to Blazor and iOS
