+++ Mapsui/Navigation/PinchTracker.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using Mapsui.Utilities;
+using System;
+using System.Collections.Generic;
 
 namespace Mapsui;
 
@@ -38,5 +40,29 @@ public class PinchTracker
         var resolutionChange = _pinchState.Radius / _previousPinchState.Radius;
 
         return new PinchManipulation(_pinchState.Center, _previousPinchState.Center, resolutionChange, rotationChange, _totalRotationDelta);
+    }
+
+    public static PinchState GetPinchState(List<MPoint> locations)
+    {
+        if (locations.Count < 2)
+            throw new ArgumentException($"Less than two locations were passed into {nameof(GetPinchState)}");
+
+        double centerX = 0;
+        double centerY = 0;
+
+        foreach (var location in locations)
+        {
+            centerX += location.X;
+            centerY += location.Y;
+        }
+
+        centerX /= locations.Count;
+        centerY /= locations.Count;
+
+        var radius = Algorithms.Distance(centerX, centerY, locations[0].X, locations[0].Y);
+
+        var angle = Math.Atan2(locations[1].Y - locations[0].Y, locations[1].X - locations[0].X) * 180.0 / Math.PI;
+
+        return new PinchState(new MPoint(centerX, centerY), radius, angle);
     }
 }
